push

+++ src/main/java/com/quant_socket/handlers/socket/sec_order_filled/SecOrderFilledHandler.java
@@ -1,12 +1,10 @@
-package com.quant_socket.handlers.socket.equities_snapshot;
+package com.quant_socket.handlers.socket.sec_order_filled;
 
-import com.quant_socket.services.EquitiesSnapshotService;
-import com.quant_socket.services.ProductService;
+import com.quant_socket.services.SecuritiesOrderFilledService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.CloseStatus;
-import org.springframework.web.socket.WebSocketMessage;
 import org.springframework.web.socket.WebSocketSession;
 import org.springframework.web.socket.handler.TextWebSocketHandler;
 
@@ -15,9 +13,9 @@ import java.net.URI;
 @Component
 @Slf4j
 @RequiredArgsConstructor
-public class EquitiesSnapshotDetailsHandler extends TextWebSocketHandler {
+public class SecOrderFilledHandler extends TextWebSocketHandler {
 
-    private final EquitiesSnapshotService service;
+    private final SecuritiesOrderFilledService service;
 
     private String[] isinCode = new String[]{};
 
@@ -32,13 +30,13 @@ public class EquitiesSnapshotDetailsHandler extends TextWebSocketHandler {
         }
     }
 
-    @Override
+    /*@Override
     public void afterConnectionClosed(WebSocketSession ws, CloseStatus status) throws Exception {
         final URI uri = ws.getUri();
         if(uri != null && uri.getQuery() != null) {
-            service.removeSession(ws, isinCode);
+            if(isinCode != null) service.removeSession(ws, isinCode);
         } else {
             ws.close();
         }
-    }
+    }*/
 }
